Altera arquivos de pastas e adiciona métodos no parser do analisador sintatico
